requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0

#pylint==2.15.3
#rust based linter
ruff==0.0.284
boto3==1.24.87
#web
fastapi==0.85.0
uvicorn==0.18.3
#others
Flask==2.0.2
gunicorn
Werkzeug==2.2.2
python-dotenv
matplotlib==3.7.2
ipykernel==6.24.0
ipywidgets==8.0.7
streamlit==1.18.1
requests==2.27.1
altair==4.0
Pillow